requirements.txt
ansimarkup==1.4.0
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
defusedxml==0.7.1
Django==3.2.7
django-cors-headers==3.8.0
django-countries==7.2.1
django-phonenumber-field==5.2.0
django-postgres-extra==2.0.2
django-softdelete==0.9.2
django-templated-mail==1.1.1
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
enforce==0.3.4
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib==3.1.1
phonenumbers==8.12.30
Pillow==8.3.1
psycopg2==2.9.1
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
structlog==20.1.0
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.6
wrapt==1.12.1
